add a small debug mode ;)

+++ webapps/news/views.py
@@ -103,8 +103,9 @@ def index(request):
     if request.POST.get('url') != "HN":
         mobi_file = save_to_file(url, dir_name=settings.KINDLE_LIVE_DIR)
         if mobi_file:
-            send_mail([settings.MY_KINDLE_MAIL,], "New documentation here", "Sent from mitnk.com", files=[mobi_file,])
-            os.remove(mobi_file)
+            if not settings.DEBUG:
+                send_mail([settings.MY_KINDLE_MAIL,], "New documentation here", "Sent from mitnk.com", files=[mobi_file,])
+                os.remove(mobi_file)
             return HttpResponse("%s Sent!" % mobi_file)
         else:
             return HttpResponse("No file generated!")
